+++ MANIFEST.in
@@ -2,5 +2,5 @@ include README.md
 include LICENSE
 include Makefile
 include pyproject.toml
-recursive-include ap_master_calibration *.py
-recursive-include ap_master_calibration/templates *.j2
+recursive-include ap_create_master *.py
+recursive-include ap_create_master/templates *.j2
