+++ ap_create_master/__init__.py
@@ -1,7 +1,7 @@
 """
 Generated By: Cursor (Claude Sonnet 4.5)
 
-ap-master-calibration: Automated calibration master generation for PixInsight.
+ap-create-master: Automated calibration master generation for PixInsight.
 """
 
 import re
