+++ ap_create_master/__main__.py
@@ -1,5 +1,5 @@
 """
-Allow running the module with python -m ap_master_calibration
+Allow running the module with python -m ap_create_master
 
 Generated By: Claude Code (Claude Sonnet 4.5)
 """
