+++ ap_create_master/script_generator.py
