+++ ap_create_master/templates/ImageIntegration_bias.j2
