+++ ap_create_master/templates/ImageIntegration_dark.j2
