+++ ap_create_master/templates/ImageIntegration_flat.j2
