+++ tests/conftest.py
@@ -1,5 +1,5 @@
 """
-Pytest configuration and shared fixtures for ap-master-calibration tests.
+Pytest configuration and shared fixtures for ap-create-master tests.
 
 Generated By: Cursor (Claude Sonnet 4.5)
 """
