+++ tests/test_calibrate_masters.py
@@ -1,5 +1,5 @@
 """
-Unit tests for ap_master_calibration.calibrate_masters module.
+Unit tests for ap_create_master.calibrate_masters module.
 
 Generated By: Cursor (Claude Sonnet 4.5)
 """
@@ -10,18 +10,18 @@ from unittest.mock import patch
 
 import pytest
 
-from ap_master_calibration import config
-from ap_master_calibration.calibrate_masters import generate_masters
+from ap_create_master import config
+from ap_create_master.calibrate_masters import generate_masters
 
 
 class TestGenerateMasters:
     """Tests for generate_masters function."""
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_master_calibration.calibrate_masters.group_files")
-    @patch("ap_master_calibration.calibrate_masters.get_group_metadata")
-    @patch("ap_master_calibration.calibrate_masters.find_matching_master_for_flat")
-    @patch("ap_master_calibration.calibrate_masters.generate_combined_script")
+    @patch("ap_create_master.calibrate_masters.group_files")
+    @patch("ap_create_master.calibrate_masters.get_group_metadata")
+    @patch("ap_create_master.calibrate_masters.find_matching_master_for_flat")
+    @patch("ap_create_master.calibrate_masters.generate_combined_script")
     def test_generates_script_for_bias_frames(
         self,
         mock_generate_script,
@@ -89,10 +89,10 @@ class TestGenerateMasters:
         mock_generate_script.assert_called_once()
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_master_calibration.calibrate_masters.group_files")
-    @patch("ap_master_calibration.calibrate_masters.get_group_metadata")
-    @patch("ap_master_calibration.calibrate_masters.find_matching_master_for_flat")
-    @patch("ap_master_calibration.calibrate_masters.generate_combined_script")
+    @patch("ap_create_master.calibrate_masters.group_files")
+    @patch("ap_create_master.calibrate_masters.get_group_metadata")
+    @patch("ap_create_master.calibrate_masters.find_matching_master_for_flat")
+    @patch("ap_create_master.calibrate_masters.generate_combined_script")
     def test_handles_no_files_gracefully(
         self,
         mock_generate_script,
@@ -116,10 +116,10 @@ class TestGenerateMasters:
         mock_generate_script.assert_not_called()
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_master_calibration.calibrate_masters.group_files")
-    @patch("ap_master_calibration.calibrate_masters.get_group_metadata")
-    @patch("ap_master_calibration.calibrate_masters.find_matching_master_for_flat")
-    @patch("ap_master_calibration.calibrate_masters.generate_combined_script")
+    @patch("ap_create_master.calibrate_masters.group_files")
+    @patch("ap_create_master.calibrate_masters.get_group_metadata")
+    @patch("ap_create_master.calibrate_masters.find_matching_master_for_flat")
+    @patch("ap_create_master.calibrate_masters.generate_combined_script")
     def test_finds_masters_for_flats(
         self,
         mock_generate_script,
@@ -199,9 +199,9 @@ class TestGenerateMasters:
         assert mock_find_master.call_count == 2
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_master_calibration.calibrate_masters.group_files")
-    @patch("ap_master_calibration.calibrate_masters.get_group_metadata")
-    @patch("ap_master_calibration.calibrate_masters.generate_combined_script")
+    @patch("ap_create_master.calibrate_masters.group_files")
+    @patch("ap_create_master.calibrate_masters.get_group_metadata")
+    @patch("ap_create_master.calibrate_masters.generate_combined_script")
     def test_uses_custom_script_output_dir(
         self,
         mock_generate_script,
@@ -285,10 +285,10 @@ class TestGenerateMasters:
         assert "dark" in captured.out.lower()
 
     @patch("ap_common.get_filtered_metadata")
-    @patch("ap_master_calibration.calibrate_masters.group_files")
-    @patch("ap_master_calibration.calibrate_masters.get_group_metadata")
-    @patch("ap_master_calibration.calibrate_masters.find_matching_master_for_flat")
-    @patch("ap_master_calibration.calibrate_masters.generate_combined_script")
+    @patch("ap_create_master.calibrate_masters.group_files")
+    @patch("ap_create_master.calibrate_masters.get_group_metadata")
+    @patch("ap_create_master.calibrate_masters.find_matching_master_for_flat")
+    @patch("ap_create_master.calibrate_masters.generate_combined_script")
     def test_handles_invalid_exposure_time_in_flats(
         self,
         mock_generate_script,
