+++ tests/test_config.py
@@ -1,10 +1,10 @@
 """
-Unit tests for ap_master_calibration.config module.
+Unit tests for ap_create_master.config module.
 
 Generated By: Cursor (Claude Sonnet 4.5)
 """
 
-from ap_master_calibration import config
+from ap_create_master import config
 
 
 class TestConfig:
