+++ tests/test_grouping.py
@@ -1,13 +1,13 @@
 """
-Unit tests for ap_master_calibration.grouping module.
+Unit tests for ap_create_master.grouping module.
 
 Generated By: Cursor (Claude Sonnet 4.5)
 """
 
 import pytest
 
-from ap_master_calibration import config
-from ap_master_calibration.grouping import (
+from ap_create_master import config
+from ap_create_master.grouping import (
     create_group_key,
     get_group_metadata,
     group_files,
