+++ tests/test_master_matching.py
@@ -1,5 +1,5 @@
 """
-Unit tests for ap_master_calibration.master_matching module.
+Unit tests for ap_create_master.master_matching module.
 
 Generated By: Cursor (Claude Sonnet 4.5)
 """
@@ -9,8 +9,8 @@ from unittest.mock import patch
 
 import pytest
 
-from ap_master_calibration import config
-from ap_master_calibration.master_matching import find_matching_master_for_flat
+from ap_create_master import config
+from ap_create_master.master_matching import find_matching_master_for_flat
 
 
 class TestFindMatchingMasterForFlat:
