+++ tests/test_script_generator.py
@@ -1,5 +1,5 @@
 """
-Unit tests for ap_master_calibration.script_generator module.
+Unit tests for ap_create_master.script_generator module.
 
 Generated By: Cursor (Claude Sonnet 4.5)
 """
@@ -8,8 +8,8 @@ from unittest.mock import patch
 
 import pytest
 
-from ap_master_calibration import config
-from ap_master_calibration.script_generator import (
+from ap_create_master import config
+from ap_create_master.script_generator import (
     escape_js_string,
     generate_combined_script,
     generate_master_filename,
